exported to separate component SnackbarMessage
